vo_direct3d: always use D3DSWAPEFFECT_DISCARD
D3DSWAPEFFECT_DISCARD will always be faster than D3DSWAPEFFECT_COPY,
since the driver is not required to clear the back buffer. Instead, the
state of the back buffer after presenting is undefined and the driver is
free to choose the fastest presentation method, whether it clobbers the
back buffer or not.